onReadNewPdf event instead of setters, deleted NavbarComponent, knowledge graph components in separate file
