+++ src/components/AppComponent/views/ReaderView.js
@@ -1,6 +1,5 @@
 import { AppView } from "./AppView.js";
 import { PdfReaderComponent } from "../../PdfReaderComponent.js";
-import { NavbarComponent } from "../../NavbarComponent.js";
 import { ThumbnailComponent } from "../../ThumbnailComponent.js";
 import { SidePageComponent } from "../../SidePageComponent.js";
 
@@ -14,7 +13,6 @@ import { SidePageComponent } from "../../SidePageComponent.js";
  */
 class ReaderView extends AppView {
   static reader = new PdfReaderComponent();
-  static navbar = new NavbarComponent();
   static thumbnail = new ThumbnailComponent();
   static sidePageComponent = new SidePageComponent();
