+++ src/components/KnowledgeGraphComponents/KnowledgeGraphComponent.js
@@ -6,13 +6,13 @@ import {
   Node,
   Link,
   GraphData,
-} from "../services/KnowledgeGraphService";
+} from "../../services/KnowledgeGraphService";
 import {
   EventHandlerService,
   PDFLEvents,
-} from "../services/EventHandlerService";
+} from "../../services/EventHandlerService";
 import { ColorLegenedComponent } from "./ColorLegenedComponent";
-import { TRANSPARENT_WHITE } from "../Constants";
+import { TRANSPARENT_WHITE } from "../../Constants";
 import { PaperInfoComponent } from "./PaperInfoComponent";
 import ForceGraph from "force-graph";
 
@@ -43,8 +43,10 @@ class KnowledgeGraphComponent {
    */
   constructor() {
     this.depth = 1;
+
     this.colorLegend = new ColorLegenedComponent();
     this.paperInfoWindow = new PaperInfoComponent(); 
+    
     this.#registerEvents();
   }
 
@@ -75,7 +77,7 @@ class KnowledgeGraphComponent {
    * @private
    * @param {int} selectedDepth new depth
    */
-  async #changeDepth(selectedDepth) {
+  #changeDepth = async (selectedDepth) => {
     try {
       EventHandlerService.publish(PDFLEvents.onShowTransparentSidePageLoader);
       await buildGraphProcedure(this.graph, selectedDepth);
@@ -97,6 +99,7 @@ class KnowledgeGraphComponent {
 
   /**
    * Displays knowledge graph.
+   * @async
    */
   displayGraph = async (depth = 1) => {
     this.components.container.classList.remove("hidden");
@@ -116,11 +119,11 @@ class KnowledgeGraphComponent {
 
   /**
    * Creates and returns custom styled force graph.
-   * 
+   * @private
    * @param {GraphData} linkedPapers nodes and links in graph
    * @returns {ForceGraph}
    */
-  #createForceGraph(linkedPapers) {
+  #createForceGraph = (linkedPapers) => {
     const currentPaperId = linkedPapers.nodes[0].id;
     const highlightNodes = new Set();
     const highlightLinks = new Set();
@@ -131,10 +134,10 @@ class KnowledgeGraphComponent {
       .nodeId("id")
       .nodeColor((node) => fieldsOfStudyToColor(node.fieldsOfStudy))
       .nodeLabel((node) => `${node.label}`)
-      .nodeVal(node => this.getNodeSize(node, currentPaperId))
+      .nodeVal(node => this.#getNodeSize(node, currentPaperId))
       .linkColor(() => TRANSPARENT_WHITE)
       .autoPauseRedraw(false)
-      .onNodeHover((node) => hoveredNode = this.displayHoveredNode(node, highlightNodes, highlightLinks))
+      .onNodeHover((node) => hoveredNode = this.#displayHoveredNode(node, highlightNodes, highlightLinks))
       .onNodeClick((node) => expandNode(node, this.graph))
       .onLinkHover((link) => this.#highlightLink(highlightNodes, highlightLinks, link))
       .linkWidth((link) => this.#getLinkWidth(highlightLinks, link))
@@ -149,16 +152,17 @@ class KnowledgeGraphComponent {
       .onEngineStop(() => this.graph.zoomToFit(500))
       .d3Force("center", null);
   }
+
   /**
    * Returns the size of node. If node's id is equal to 
    * current paper id that node will be bigger than the rest
    * of nodes.
-   * 
+   * @private
    * @param {Node} node node being processed
    * @param {string} currentPaperId id of paper being read
    * @returns {int} node size
    */
-  getNodeSize(node, currentPaperId) {
+  #getNodeSize = (node, currentPaperId) => {
     return Math.pow(node.id === currentPaperId ? 2 : 1, 2);
   }
 
@@ -166,13 +170,13 @@ class KnowledgeGraphComponent {
    * Displays node that is being hovered. When node is hovered it is
    * highlighted, as well as it's connected nodes and links. Paper
    * information popup is displayed also for hovered node.
-   * 
+   * @private
    * @param {Node} node node being processed
    * @param {Set<Node>} highlightNodes currently highlighted nodes
    * @param {Set<Link>} highlightLinks currently highlighted links
    * @returns {Node} styled ForceGraph's node that is being hovered
    */
-  displayHoveredNode(node, highlightNodes, highlightLinks) {
+  #displayHoveredNode = (node, highlightNodes, highlightLinks) => {
     this.paperInfoWindow.displayPaperInfo(node);
     return this.#highlightConnectedNodes(highlightNodes, highlightLinks, node);
   }
@@ -182,13 +186,13 @@ class KnowledgeGraphComponent {
    * that are connected to that node over links. Clears old
    * highlighted nodes and links sets and finds new links and nodes
    * connected to newly hovered node.
-   *
+   * @private
    * @param {Set<Node>} highlightNodes currently highlighted nodes
    * @param {Set<Link>} highlightLinks currently highlighted links
    * @param {Node} node node currently being hovered
    * @returns {Node} styled ForceGraph's node that is being hovered
    */
-  #highlightConnectedNodes(highlightNodes, highlightLinks, node) {
+  #highlightConnectedNodes = (highlightNodes, highlightLinks, node) => {
     highlightNodes.clear();
     highlightLinks.clear();
 
@@ -216,13 +220,13 @@ class KnowledgeGraphComponent {
    * is checked and if source/target of that link is current link
    * then that link's target/source node if found in current graph
    * data and finally all such nodes are returned in array.
-   *
+   * @private
    * @param {Array<Link>} nodeLinks all links connected to hovered node
    * @param {Node} node node that is being hovered over
    * @param {GraphData} graphData current graph data
    * @returns {Array<Node>} nodes connected to node over links
    */
-  #findHighlightedNodes(nodeLinks, node, graphData) {
+  #findHighlightedNodes = (nodeLinks, node, graphData) => {
     return nodeLinks.map((l) => {
       if (l.source.id == node.id)
         return graphData.nodes.find((n) => n.id == l.target.id);
@@ -234,12 +238,12 @@ class KnowledgeGraphComponent {
    * Returns array of  links that are connected to given node.
    * Link is connected to node if link's target or source id is
    * equal to node id.
-   *
+   * @private
    * @param {GraphData} graphData current graph data
    * @param {Node} node node that is being hovered over
    * @returns {Array<Link>} links connected to node
    */
-  #findNodesLinks(graphData, node) {
+  #findNodesLinks = (graphData, node) => {
     return graphData.links.filter(
       (l) => l.source.id == node.id || l.target.id == node.id
     );
@@ -248,13 +252,13 @@ class KnowledgeGraphComponent {
   /**
    * Highlights a link. Adds link to highlighted links set and nodes
    * it connects to highlighted nodes set.
-   *
+   * @private
    * @param {Set<Node>} highlightNodes  set of highlighted nodes
    * @param {Set<Link>} highlightLinks set of highlighted links
    * @param {Link} link link being hovered over
    * @returns {Link} styled ForceGraph's link
    */
-  #highlightLink(highlightNodes, highlightLinks, link) {
+  #highlightLink = (highlightNodes, highlightLinks, link) => {
     highlightNodes.clear();
     highlightLinks.clear();
 
@@ -270,73 +274,72 @@ class KnowledgeGraphComponent {
   /**
    * Returns link width depending if link is highlighted. If it is
    * highlighted returns 5 and 1 otherwise.
-   *
+   * @private
    * @param {Set<Link>} highlightLinks set of highlighted links
    * @param {Link} link link being processed
    * @returns {int} link width
    */
-  #getLinkWidth(highlightLinks, link) {
+  #getLinkWidth = (highlightLinks, link) => {
     return this.#isLinkHighlighted(highlightLinks, link) ? 5 : 1;
   }
 
   /**
    * Returns length of directional arrow depending if link is highlighted.
    * If link is hovered returns 16 and 8 otherwise.
-   *
+   * @private
    * @param {Set<Link>} highlightLinks set of highlighted links
    * @param {Link} link link being processed
    * @returns {int} directional arrow length
    */
-  #getArrowLength(highlightLinks, link) {
+  #getArrowLength = (highlightLinks, link) => {
     return this.#isLinkHighlighted(highlightLinks, link) ? 16 : 8;
   }
 
   /**
    * Returns width of directional particle depending if link is highlighted.
    * If link is hovered returns 4 and 2 otherwise.
-   *
+   * @private
    * @param {Set<Link>} highlightLinks set of highlighted links
    * @param {Link} link link being processed
    * @returns {int} directional particle width
    */
-  #getParticleWidth(highlightLinks, link) {
+  #getParticleWidth = (highlightLinks, link) => {
     return this.#isLinkHighlighted(highlightLinks, link) ? 4 : 2;
   }
 
   /**
    * Returns true if link is in highlighted links.
-   *
+   * @private
    * @param {Set<Link>} highlightLinks set of highlighted links
    * @param {Link} link link in graph
    * @returns {boolean}
    */
-  #isLinkHighlighted(highlightLinks, link) {
+  #isLinkHighlighted = (highlightLinks, link) => {
     return highlightLinks.has(link);
   }
 
   /**
    * Returns node mode depending if highlighted. If node is
    * highlighted returns "before" and undefined otherwise.
-   *
+   * @private
    * @param {Set<Node>} highlightNodes set of highlighted nodes
    * @param {Node} node node being processed
    * @returns {string} node mode
    */
-  #getNodeMode(highlightNodes, node) {
+  #getNodeMode = (highlightNodes, node) => {
     return highlightNodes.has(node) ? "before" : undefined;
   }
 
   /**
    * Sets style of ring around node depending if node is highlighted or not.
-   *
+   * @private
    * @param {Node} hoveredNode last hovered node
    * @param {Node} node node being processed
    * @param {Object} ctx canvas context of node
    * @returns {(Node, Object)} ForceGraph's styled node and context
    */
-  #displayHighlightedNode(hoveredNode, currentPaperId, node, ctx) {
+  #displayHighlightedNode = (hoveredNode, currentPaperId, node, ctx) => {
     const nodeRadius = 4;
-    // add ring just for highlighted nodes
     ctx.beginPath();
     ctx.arc(node.x, node.y, nodeRadius * (node.id === currentPaperId ? 2.4 : 1.4), 0, 2 * Math.PI, false)
     ctx.fillStyle = hoveredNode && node.id === hoveredNode.id ? "red" : "orange";
@@ -348,14 +351,14 @@ class KnowledgeGraphComponent {
    * Returns true if this component is displayed in side window and false otherwise.
    * @returns {boolean}
    */
-  isOpened(){
+  isOpened = () => {
     return !this.components.container.classList.contains("hidden");
   }
 
   /**
    * Hides this whole component.
    */
-  hide() {
+  hide = () => {
     this.components.container.classList.add("hidden");
   }
