+++ src/components/ToolbarComponent.js
@@ -4,29 +4,39 @@ import {
 } from "../services/EventHandlerService";
 import { PaginationComponent } from "./PaginationComponent";
 import { ZoomComponent } from "./ZoomComponent";
+import { POPUP_DISAPPEAR_TIMEOUT } from "../Constants";
+import { readFile } from "../services/FileUploadService";
 
 /**
  * Component that serves as a placehodler for all elements in a toolbar. Manages functionality of it's
  * elements, as well as linking components to their methods.
  *
  * @property {Object} components object that holds all DOM elements within this component
- * @property {HTMLElement} components.toolbar element that represents this whole component
  * @property {HTMLElement} components.fullScreen full screen button
  * @property {HTMLElement} components.graphMakerBtn button that generates knowledge graph
  * @property {HTMLElement} components.body body of HTML document
  * @property {HTMLElement} components.thumbnailBtn button that opens the thumbnail
  * @property {PaginationComponent} paginationComponent pagination component
  * @property {ZoomComponent} zoomComponent zoom component
+ * 
  * @property {HTMLElement} components.summaryKeyBtn button that open and close the sidepage
+ * @property {HTMLElement} components.openNew button to click for upload a new pdf from the reader component
+ * @property {HTMLElement} components.loader loader for showing the pdf uploaded
+ * @property {HTMLElement} components.errorMessage error message for a wrong pdf uploaded from the pdf reader component
  */
 class ToolbarComponent {
   components = {
-    toolbar: document.querySelector("#sidenav"),
     fullScreen: document.querySelector("#full-screen"),
     graphMakerBtn: document.querySelector("#graph-maker"),
     body: document.querySelector("body"),
     summaryKeyBtn: document.querySelector("#summary-maker"),
     thumbnailBtn: document.querySelector("#pages-sidebar"),
+
+    openNew: document.querySelector("#open-new-pdf"),
+    loader: document.querySelector("#loader"),
+    errorMessage: document.querySelector(
+      "#message-wrong-type-fileupload-reader"
+    ),
   };
 
   /**
@@ -66,6 +76,19 @@ class ToolbarComponent {
       (functionalKeys, key, code) => {
         if (functionalKeys.alt && code === "KeyL") {
           this.#showFullScreen();
+        }
+      }
+    );
+
+    this.components.openNew.addEventListener("input", this.#onNewFile);
+    EventHandlerService.subscribe(
+      PDFLEvents.onKeyboardKeyDown,
+      (functionalKeys, key) => {
+        if (!functionalKeys.ctrl) {
+          return;
+        }
+        if (key === "u") {
+          this.components.openNew.click();
         }
       }
     );
@@ -150,5 +173,49 @@ class ToolbarComponent {
     this.paginationComponent.setCurrentPage(1);
     this.zoomComponent.setZoom(1);
   };
+
+
+    /**
+   * Call back for checking file format and read it. If the file is not a pdf an error message is displayed
+   * @private
+   */
+    #onNewFile = (event) => {
+      if(!event.target.files[0]) return;
+      if (event.target.files[0].type === "application/pdf") {
+        this.#showLoader();
+        this.#hideErrorMessage();
+        readFile(event.target.files[0]);
+      } else {
+        this.#showErrorMessage();
+      }
+    };
+  
+    /**
+     * Hide error message
+     * @private
+     */
+    #hideErrorMessage = () => {
+      this.components.errorMessage.classList.add("hidden");
+    };
+  
+    /**
+     * Show error message
+     * @private
+     */
+    #showErrorMessage = () => {
+      clearTimeout(messageErrorTimeOut);
+      this.components.errorMessage.classList.remove("hidden");
+      var messageErrorTimeOut = setTimeout(() => {
+        this.hideErrorMessage();
+      }, POPUP_DISAPPEAR_TIMEOUT);
+    };
+  
+    /**
+     * Show loader
+     * @private
+     */
+    #showLoader = () => {
+      this.components.loader.classList.remove("hidden");
+    };
 }
 export { ToolbarComponent };
